backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
pymongo==3.12.0
numpy==1.21.0
pandas==1.3.0
tensorflow==2.6.0
torch==1.9.0
scikit-learn==0.24.2
opencv-python==4.5.3.56
earthengine-api==0.1.270
python-dotenv==0.19.0
requests==2.26.0
gunicorn==20.1.0
geopandas==0.9.0
rasterio==1.2.10
matplotlib==3.4.2
seaborn==0.11.2 